requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
coreapi==2.3.0
coreschema==0.0.4
Django==5.1.5
django-cors-headers==4.6.0
django-rest-framework-docs==0.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonpickle==4.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
packaging==24.2
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
